+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "vscode-websearch" extension will be documented in th
 - Initial release
 - Added:
   - Support for mac.
+  - Multiple search engine support.
 - Changed:
 - Fixed:
 - Deprecated: